factor out requirements code
